Bugfix Plugin Action Links
